More work on bear trap.

+++ include/extern.h
@@ -3178,9 +3178,9 @@ E struct monst *FDECL(animate_statue,
                       (struct obj *, XCHAR_P, XCHAR_P, int, int *));
 E struct monst *FDECL(activate_statue_trap,
                       (struct trap *, XCHAR_P, XCHAR_P, BOOLEAN_P));
-E void FDECL(set_utrap, (unsigned, unsigned));
+E void FDECL(set_utrap, (unsigned long, uchar));
 E void FDECL(reset_utrap, (BOOLEAN_P));
-E void FDECL(dotrap, (struct trap *, unsigned));
+E void FDECL(dotrap, (struct trap *, unsigned short));
 E void FDECL(seetrap, (struct trap *));
 E void FDECL(feeltrap, (struct trap *));
 E int FDECL(mintrap, (struct monst *));
