+++ src/hack.c
@@ -2498,7 +2498,7 @@ boolean pick;
 
     struct monst *mtmp;
     struct trap *trap = t_at(u.ux, u.uy);
-    int trapflag = iflags.failing_untrap ? FAILEDUNTRAP : 0;
+    unsigned short trapflag = iflags.failing_untrap ? FAILEDUNTRAP : 0;
 
     /* prevent recursion from affecting the hero all over again
        [hero poly'd to iron golem enters water here, drown() inflicts
